requirements.txt
flask
transformers
torch
datasets
accelerate
pydub
soundfile
APScheduler
python-dotenv
requests
pyngrok